// src/app/models/classes/customer.ts
export class Customer {
  id: number;
  name: string;
  email?: string;
  phone?: string;
  pricelist_id?: number;
  pricelist_name?: string;
  // address?: string;
  // last_visit?: Date;
  // first_visit?: Date;
  // favorite_dishes: string[];
  // birthday?: Date;
  // feedback?: string[];
  // reservation_history: Reservation[];
  // payment_preference?: string;
  // status?: string;
}